refactor: move code into hooks
